# annotation/documentation/openapi.yaml
openapi: 3.0.2
info:
  title: Badgerdoc Annotation
  version: 0.1.5
paths:
...
  /tasks/{task_id}:
    get:
      tags:
      - Manual Annotation Tasks
      summary: Get manual annotation task by id.
      operationId: get_task_tasks__task_id__get
      parameters:
      - required: true
        schema:
          title: Task Id
          type: integer
        example: 1
        name: task_id
        in: path
      - required: true
        schema:
          title: X-Current-Tenant
          type: string
        example: test
        name: X-Current-Tenant
        in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpandedManualAnnotationTaskSchema'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionErrorSchema'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - TenantDependencyDocs: []
    delete:
      tags:
      - Manual Annotation Tasks
      summary: Delete task by id.
      operationId: delete_task_tasks__task_id__delete
      parameters:
      - required: true
        schema:
          title: Task Id
          type: integer
        example: 3
        name: task_id
        in: path
      - required: true
        schema:
          title: X-Current-Tenant
          type: string
        example: test
        name: X-Current-Tenant
        in: header
      responses:
        '204':
          description: Successful Response
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionErrorSchema'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - TenantDependencyDocs: []
    patch:
      tags:
      - Manual Annotation Tasks
      summary: Update task by id.
      description: |-
        Takes not required fields file_id, pages, job_id,
        user_id or (and) is_validation. By task`s id
        in path, updates task with tasks file and returns updated task.
      operationId: update_task_tasks__task_id__patch
      parameters:
      - required: true
        schema:
          title: Task Id
          type: integer
        example: 5
        name: task_id
        in: path
      - required: true
        schema:
          title: X-Current-Tenant
          type: string
        example: test
        name: X-Current-Tenant
        in: header
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskPatchSchema'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManualAnnotationTaskSchema'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionErrorSchema'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestErrorSchema'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - TenantDependencyDocs: []
...
components:
  schemas:
...
    BadRequestErrorSchema:
      title: BadRequestErrorSchema
      required:
      - detail
      type: object
      properties:
        detail:
          title: Detail
          type: string
      example:
        detail: Bad request.
...
    ConnectionErrorSchema:
      title: ConnectionErrorSchema
      required:
      - detail
      type: object
      properties:
        detail:
          title: Detail
          type: string
      example:
        detail: 'Error: Connection error.'
...
    ExpandedManualAnnotationTaskSchema:
      title: ExpandedManualAnnotationTaskSchema
      required:
      - id
      - pages
      - user_id
      - file
      - job
      type: object
      properties:
        id:
          title: Id
          type: integer
          example: 4
        status:
          allOf:
          - $ref: '#/components/schemas/TaskStatusEnumSchema'
          default: Pending
          example: Pending
        pages:
          title: Pages
          minimum: 1.0
          minItems: 1
          uniqueItems: true
          type: array
          items:
            minimum: 1.0
            type: integer
          example:
          - 1
          - 2
          - 3
        user_id:
          title: User Id
          type: string
          format: uuid
          example: 3082242e-15e3-4e18-aad0-e3bf182b8550
        is_validation:
          title: Is Validation
          type: boolean
          default: false
          example: false
        deadline:
          title: Deadline
          type: string
          format: date-time
          example: '2021-10-19 01:01:01'
        file:
          $ref: '#/components/schemas/NameSchema'
        job:
          $ref: '#/components/schemas/NameSchema'
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
...
    ManualAnnotationTaskSchema:
      title: ManualAnnotationTaskSchema
      required:
      - id
      - file_id
      - pages
      - job_id
      - user_id
      type: object
      properties:
        id:
          title: Id
          type: integer
          example: 4
        status:
          allOf:
          - $ref: '#/components/schemas/TaskStatusEnumSchema'
          default: Pending
          example: Pending
        file_id:
          title: File Id
          type: integer
          example: 2
        pages:
          title: Pages
          minimum: 1.0
          minItems: 1
          uniqueItems: true
          type: array
          items:
            minimum: 1.0
            type: integer
          example:
          - 1
          - 2
          - 3
        job_id:
          title: Job Id
          type: integer
          example: 3
        user_id:
          title: User Id
          type: string
          format: uuid
          example: 4e9c5839-f63b-49c8-b918-614b87813e53
        is_validation:
          title: Is Validation
          type: boolean
          default: false
          example: false
        deadline:
          title: Deadline
          type: string
          format: date-time
          example: '2021-10-19 01:01:01'
    NameSchema:
      title: NameSchema
      required:
      - id
      type: object
      properties:
        id:
          title: Id
          type: integer
          example: 1
        name:
          title: Name
          type: string
          example: NameOfJobOrFile
    NotFoundErrorSchema:
      title: NotFoundErrorSchema
      required:
      - detail
      type: object
      properties:
        detail:
          title: Detail
          type: string
      example:
        detail: Resource was not found.
...
    TaskPatchSchema:
      title: TaskPatchSchema
      type: object
      properties:
        file_id:
          title: File Id
          type: integer
          example: 2
        pages:
          title: Pages
          minimum: 1.0
          minItems: 1
          uniqueItems: true
          type: array
          items:
            minimum: 1.0
            type: integer
          example:
          - 1
          - 2
          - 3
        job_id:
          title: Job Id
          type: integer
          example: 3
        user_id:
          title: User Id
          type: string
          format: uuid
          example: 4e9c5839-f63b-49c8-b918-614b87813e53
        is_validation:
          title: Is Validation
          type: boolean
          example: false
        deadline:
          title: Deadline
          type: string
          format: date-time
          example: '2021-10-19 01:01:01'
    TaskStatusEnumSchema:
      title: TaskStatusEnumSchema
      enum:
      - Pending
      - Ready
      - In Progress
      - Finished
      type: string
      description: An enumeration.
...
    ValidationError:
      title: ValidationError
      required:
      - loc
      - msg
      - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            type: string
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
...
  securitySchemes:
    TenantDependencyDocs:
      type: http
      scheme: bearer
...